docs(supersync): correct the plan-cache and aggregate claims; match Int in tests
Round-4 review. Four inaccuracies, all in text a contributor would rely on.

- plan_cache_mode=auto does NOT unconditionally switch to generic after 5
  executions: it compares the generic cost against the average custom cost
  and may keep using custom plans indefinitely. Stated as an observation for
  THIS statement (custom_plans=5, generic_plans=15 on production) rather than
  a universal rule, in conflict.ts, the plan spec and vector-clocks.md.

  Note: commit fdf0f62011 claimed this fix for conflict.ts. It did not
  contain it - the edit was lost and the commit message was not verified
  against the diff. Applied here for real.

- Prisma's bare aggregate() cannot return zero rows; it returns one object
  with _max.serverSeq null. Only GROUP BY can produce zero rows. The fold
  comment claimed otherwise.

- The test schema declared server_seq bigint while production maps it as
  Prisma Int. MAX() therefore returned bigint and the shim's Number()
  coercion hid it, so dropping the ::int cast from the production SQL would
  have stayed green. Test schema now matches production.

- detectConflictForEntity's JSDoc still described the path as using Prisma's
  typed model API; its array branch is raw SQL. vector-clocks.md's flow
  summary likewise still described a single findFirst.

+++ packages/super-sync-server/tests/conflict-entity-lookup-plan.pglite.spec.ts
@@ -19,10 +19,12 @@ import type { Operation } from '../src/sync/sync.types';
  * fail here instead of passing against a stale copy.
  *
  * MEASURE WITH `force_generic_plan`, NEVER WITH LITERALS. Prisma sends parameterized
- * prepared statements, and under the default `auto` Postgres builds CUSTOM plans for the
- * first ~5 executions before settling on a GENERIC one that cannot see parameter values.
- * A hot path lives in that steady state, so generic is the mode that matters — but note
- * production does receive custom plans too, and this file does not cover them (a
+ * prepared statements; under the default `auto` Postgres plans the first ~5 executions as
+ * CUSTOM, then compares the generic cost against the average custom cost and MAY switch to
+ * a generic plan — a cost comparison, not an automatic switch, so a statement can stay on
+ * custom plans indefinitely. THIS one was observed going generic on production, and a
+ * generic plan cannot see parameter values, so that is the mode this file covers.
+ * Production also serves custom plans and this file does NOT cover them (a
  * custom-plan-only regression is possible; see the rejected `server_seq >` narrowing in
  * conflict.ts). `EXPLAIN` with literal constants is a third thing again, and is the trap:
  * this file once tested that way and the blind spot passed two designs that were
@@ -86,7 +88,10 @@ const CREATE_TABLE = `
     id             text PRIMARY KEY,
     user_id        integer NOT NULL,
     client_id      text NOT NULL,
-    server_seq     bigint NOT NULL,
+    -- integer, NOT bigint: production maps serverSeq as Prisma Int. Declaring bigint
+    -- here made MAX() return bigint, which the shim's Number() coercion then hid, so
+    -- dropping the ::int cast from the production SQL would have stayed green.
+    server_seq     integer NOT NULL,
     action_type    text NOT NULL,
     entity_type    text NOT NULL,
     entity_id      text,
